removed all references to protected variable quality from tests
