+++ src/main/java/org/agilegrenoble/objectcalisthenics/ItemForge.java
@@ -13,8 +13,7 @@ import org.agilegrenoble.objectcalisthenics.quality.Quality;
 public class ItemForge {
 
     public Item anAgedBrie_thatImprovesWithTime(int daysUntilBestBeforeDate, int startQuality) {
-        return new NormalItem("Aged Brie", 
-                new ImproveWithAge(daysUntilBestBeforeDate), new Between0and50Quality(startQuality));
+        return anAgedBrie_thatImprovesWithTime(daysUntilBestBeforeDate, new Between0and50Quality(startQuality));
     }
 
     public Item anAgedBrie_thatImprovesWithTime(int daysUntilBestBeforeDate, Quality quality) {
@@ -27,21 +26,19 @@ public class ItemForge {
     }
 
     public Item anItem_thatDecaysWithTime(String name, int daysUntilBestBeforeDate, int startQuality) {
-        return new NormalItem(name, 
-                new DecayWithAge(daysUntilBestBeforeDate), new Between0and50Quality(startQuality));
-    }
-
-    public Item aBackstagePass_thatImprovesUntilTheConcertDate(int daysUntilConcert, int startQuality) {
-        return new NormalItem("Backstage passes to a TAFKAL80ETC concert", 
-                new BackStagePassAgeing(daysUntilConcert), new Between0and50Quality(startQuality));
-    }
-    public BackStagePass aBackstagePass_thatImprovesUntilTheConcertDate(int daysUntilConcert, Quality quality) {
-        return new BackStagePass(quality, new BackStagePassAgeing(daysUntilConcert));
+        return anItem_thatDecaysWithTime(name, daysUntilBestBeforeDate, new Between0and50Quality(startQuality));
     }
 
     public Item anItem_thatDecaysWithTime(String itemName, int startAge, Quality quality) {
         return new NormalItem(itemName, 
                 new DecayWithAge(startAge), quality);
+    }
+    
+    public BackStagePass aBackstagePass_thatImprovesUntilTheConcertDate(int daysUntilConcert, int startQuality) {
+        return aBackstagePass_thatImprovesUntilTheConcertDate(daysUntilConcert, new Between0and50Quality(startQuality));
+    }
+    public BackStagePass aBackstagePass_thatImprovesUntilTheConcertDate(int daysUntilConcert, Quality quality) {
+        return new BackStagePass(quality, new BackStagePassAgeing(daysUntilConcert));
     }
