+++ src/main/java/org/agilegrenoble/objectcalisthenics/items/Item.java
@@ -16,10 +16,6 @@ public abstract class Item {
 
     public abstract void updateQuality();
 
-    public int getQuality() {
-        return quality.getQuality();
-    }
-
     protected int getDayCountDown() {
         return ageing.getDayCountDown();
     }
