+++ src/test/java/org/agilegrenoble/objectcalisthenics/quality/TestBetween0and50QualityTest.java
@@ -5,6 +5,7 @@ import static org.fest.assertions.api.Assertions.assertThat;
 import java.util.Random;
 
 import org.agilegrenoble.objectcalisthenics.items.Item;
+import org.fest.assertions.api.Assertions;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -15,6 +16,52 @@ public class TestBetween0and50QualityTest {
     @Before
     public void setup() {
         random = new Random();
+    }
+
+    @Test
+    public void increaseQuality_ShouldIncreaseQualityByTheGivenValue(){
+        // Given
+        Quality quality = quality(6);
+        
+        int value = random.nextInt();
+        
+        // When
+        quality.increase(value);
+        
+        // Then
+        Quality expectedQuality = quality(6 + value);
+        Assertions.assertThat(quality).isEqualTo(expectedQuality);
+    }
+    
+    @Test
+    public void decreaseQuality_ShouldDecreaseQualityByTheGivenValue(){
+        // Given
+        int startQuality = random.nextInt(51);
+        Quality quality = quality(startQuality);
+        
+        int value = random.nextInt();
+        
+        // When
+        quality.decrease(value);
+        
+        // Then
+        int expectedQuality = startQuality - value;
+        Assertions.assertThat(quality).isEqualTo(quality(expectedQuality));
+    }
+    
+    @Test
+    public void resetQualityToZero_ShouldResetQualityToZero()
+    {
+        // Given
+        int startQuality = random.nextInt();
+        Quality quality = quality(startQuality);
+
+        
+        // When
+        quality.resetToZero();
+        
+        // Then
+        Assertions.assertThat(quality).isEqualTo(quality(0));
     }
 
     @Test
